+++ src/setupTests.ts
@@ -1,6 +1,11 @@
 import "@testing-library/jest-dom";
 import "jest-canvas-mock";
 
+jest.mock("nanoid", () => {
+  return {
+    nanoid: jest.fn(() => "test-id"),
+  };
+});
 // ReactDOM is located inside index.tsx file
 // as a result, we need a place for it to render into
 const element = document.createElement("div");